+++ react-library/src/components/stencil-generated/index.ts
@@ -12,7 +12,7 @@ export const FormControlSelect = /*@__PURE__*/createReactComponent<JSX.FormContr
 export const FormControlText = /*@__PURE__*/createReactComponent<JSX.FormControlText, HTMLFormControlTextElement>('form-control-text');
 export const JustifiBankAccountForm = /*@__PURE__*/createReactComponent<JSX.JustifiBankAccountForm, HTMLJustifiBankAccountFormElement>('justifi-bank-account-form');
 export const JustifiBillingForm = /*@__PURE__*/createReactComponent<JSX.JustifiBillingForm, HTMLJustifiBillingFormElement>('justifi-billing-form');
-export const JustifiBusinessAddress = /*@__PURE__*/createReactComponent<JSX.JustifiBusinessAddress, HTMLJustifiBusinessAddressElement>('justifi-business-address');
+export const JustifiBusinessAddressForm = /*@__PURE__*/createReactComponent<JSX.JustifiBusinessAddressForm, HTMLJustifiBusinessAddressFormElement>('justifi-business-address-form');
 export const JustifiBusinessForm = /*@__PURE__*/createReactComponent<JSX.JustifiBusinessForm, HTMLJustifiBusinessFormElement>('justifi-business-form');
 export const JustifiCardForm = /*@__PURE__*/createReactComponent<JSX.JustifiCardForm, HTMLJustifiCardFormElement>('justifi-card-form');
 export const JustifiPaymentForm = /*@__PURE__*/createReactComponent<JSX.JustifiPaymentForm, HTMLJustifiPaymentFormElement>('justifi-payment-form');
