+++ stencil-library/src/components.d.ts
@@ -8,12 +8,12 @@ import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { CreatePaymentMethodResponse } from "./components/payment-method-form/payment-method-responses";
 import { BillingFormFields } from "./components/billing-form/billing-form-schema";
 import { ValidationError } from "yup";
-import { BusinessAddressFormFields } from "./components/business-form/business-address/business-address-schema";
+import { BusinessAddressFormFields } from "./components/business-form/business-address/business-address-form-schema";
 import { PaymentMethodTypes } from "./api";
 export { CreatePaymentMethodResponse } from "./components/payment-method-form/payment-method-responses";
 export { BillingFormFields } from "./components/billing-form/billing-form-schema";
 export { ValidationError } from "yup";
-export { BusinessAddressFormFields } from "./components/business-form/business-address/business-address-schema";
+export { BusinessAddressFormFields } from "./components/business-form/business-address/business-address-form-schema";
 export { PaymentMethodTypes } from "./api";
 export namespace Components {
     interface FormControlSelect {
@@ -77,7 +77,7 @@ export namespace Components {
          */
         "validate": () => Promise<{ isValid: boolean; }>;
     }
-    interface JustifiBusinessAddress {
+    interface JustifiBusinessAddressForm {
         "submit": () => Promise<{ isValid: boolean; values: BusinessAddressFormFields; }>;
     }
     /**
@@ -227,11 +227,11 @@ declare global {
         prototype: HTMLJustifiBillingFormElement;
         new (): HTMLJustifiBillingFormElement;
     };
-    interface HTMLJustifiBusinessAddressElement extends Components.JustifiBusinessAddress, HTMLStencilElement {
+    interface HTMLJustifiBusinessAddressFormElement extends Components.JustifiBusinessAddressForm, HTMLStencilElement {
     }
-    var HTMLJustifiBusinessAddressElement: {
-        prototype: HTMLJustifiBusinessAddressElement;
-        new (): HTMLJustifiBusinessAddressElement;
+    var HTMLJustifiBusinessAddressFormElement: {
+        prototype: HTMLJustifiBusinessAddressFormElement;
+        new (): HTMLJustifiBusinessAddressFormElement;
     };
     /**
      * @exportedPart label: Label for inputs
@@ -302,7 +302,7 @@ declare global {
         "form-control-text": HTMLFormControlTextElement;
         "justifi-bank-account-form": HTMLJustifiBankAccountFormElement;
         "justifi-billing-form": HTMLJustifiBillingFormElement;
-        "justifi-business-address": HTMLJustifiBusinessAddressElement;
+        "justifi-business-address-form": HTMLJustifiBusinessAddressFormElement;
         "justifi-business-form": HTMLJustifiBusinessFormElement;
         "justifi-card-form": HTMLJustifiCardFormElement;
         "justifi-payment-form": HTMLJustifiPaymentFormElement;
@@ -364,7 +364,7 @@ declare namespace LocalJSX {
          */
         "legend"?: string;
     }
-    interface JustifiBusinessAddress {
+    interface JustifiBusinessAddressForm {
     }
     /**
      * @exportedPart label: Label for inputs
@@ -462,7 +462,7 @@ declare namespace LocalJSX {
         "form-control-text": FormControlText;
         "justifi-bank-account-form": JustifiBankAccountForm;
         "justifi-billing-form": JustifiBillingForm;
-        "justifi-business-address": JustifiBusinessAddress;
+        "justifi-business-address-form": JustifiBusinessAddressForm;
         "justifi-business-form": JustifiBusinessForm;
         "justifi-card-form": JustifiCardForm;
         "justifi-payment-form": JustifiPaymentForm;
@@ -486,7 +486,7 @@ declare module "@stencil/core" {
              * @exportedPart input-invalid: Invalid state for inputs
              */
             "justifi-billing-form": LocalJSX.JustifiBillingForm & JSXBase.HTMLAttributes<HTMLJustifiBillingFormElement>;
-            "justifi-business-address": LocalJSX.JustifiBusinessAddress & JSXBase.HTMLAttributes<HTMLJustifiBusinessAddressElement>;
+            "justifi-business-address-form": LocalJSX.JustifiBusinessAddressForm & JSXBase.HTMLAttributes<HTMLJustifiBusinessAddressFormElement>;
             /**
              * @exportedPart label: Label for inputs
              * @exportedPart input: The input fields
