+++ stencil-library/src/components/business-form/business-address/readme.md
@@ -28,9 +28,9 @@ Type: `Promise<{ isValid: boolean; values: BusinessAddressFormFields; }>`
 ### Graph
 ```mermaid
 graph TD;
-  justifi-business-address --> text-input
-  justifi-business-address --> select-input
-  style justifi-business-address fill:#f9f,stroke:#333,stroke-width:4px
+  justifi-business-address-form --> text-input
+  justifi-business-address-form --> select-input
+  style justifi-business-address-form fill:#f9f,stroke:#333,stroke-width:4px
 ```
 
 ----------------------------------------------
