+++ stencil-library/src/components/select-input/readme.md
@@ -33,15 +33,13 @@
 ### Used by
 
  - [justifi-billing-form](../billing-form)
- - [justifi-business-address](../business-address)
- - [justifi-business-info](../business-info)
+ - [justifi-business-address-form](../business-form/business-address)
 
 ### Graph
 ```mermaid
 graph TD;
   justifi-billing-form --> select-input
-  justifi-business-address --> select-input
-  justifi-business-info --> select-input
+  justifi-business-address-form --> select-input
   style select-input fill:#f9f,stroke:#333,stroke-width:4px
 ```
