+++ stencil-library/src/components/text-input/readme.md
@@ -32,15 +32,13 @@
 ### Used by
 
  - [justifi-billing-form](../billing-form)
- - [justifi-business-address](../business-address)
- - [justifi-business-info](../business-info)
+ - [justifi-business-address-form](../business-form/business-address)
 
 ### Graph
 ```mermaid
 graph TD;
   justifi-billing-form --> text-input
-  justifi-business-address --> text-input
-  justifi-business-info --> text-input
+  justifi-business-address-form --> text-input
   style text-input fill:#f9f,stroke:#333,stroke-width:4px
 ```
